fix: move generate token to helper folder
